Add a utility class to normalize type names
PHP recognizes "bool" and "boolean", "int" and "integer", and "float" and
"double". We should allow either, but normalize to one or the other. I've chosen
the normalization

    bool -> boolean
    int -> integer
    float -> double

to match the gettype() function's return value, which we could perhaps use in
the future to validate that a provider returns a value of the requested type.