Restructure exception types
I've moved ProviderNotFoundException into Phodam\Store because it's not a
provider-thrown exception, but rather thrown by the code managing providers --
namely the ProviderStore. (At least it will be soon.)

I've also removed the $type property, and so also the custom constructor. This
exception represents -- at least in my mind -- programmer error. Therefore it
shouldn't be caught, and so therefore tracking metadata like the type argument
is irrelevant. And if it *is* caught, the caller ought to know the type they
passed to the relevant method.

Finally, I've added a base exception type, PhodamException. This is just for
organization -- so our exceptions can have a common hierarchy.